refactor(api): add labware origin calculation context (#19253)
Works toward EXEC-1690

When calculating the stackup origin to the labware origin, the finalized labware origin is dependent on the context in which we utilize the labware origin. For example, when we drop labware in a slot, the labware origin offset is typically "lower" than the default offset, because we want to press the labware into the slot. We need to account for these differences, and we want to do so in our labware origin math API.

Currently, there are three identifiable contexts after auditing the code:

- Pipetting context. The "default" context in the sense that there are no special offsets that require consideration. It's utilized typically when we move from the labware origin to a well for pipetting purposes.
- Gripper pick up labware context. When we pick up labware, we apply special gripper pick up offsets.
- Gripper drop labware context. When we drop labware, we apply special gripper drop offsets.

None of the nuances of those contexts are implemented in this commit.