fix(protocol-designer,-step-generation): fix bad move labware state updates from off-to-on deck
This PR updates the logic accessing an offdeck labware in `getFullStackFromLabwares`. The bug stemmed from the fact that all off-deck labware stacks specify `"offDeck"` as their "slot" (or bottom-most stack element). Therefore, when searching for a labware stack by slot, we return the first encountered off-deck stack. To solve this issue, we provide a labware ID as an override for off-deck stacks.

Closes AUTH-2200