feat(api): Implement cancelling previous run
To cancel previous tasks, we keep a queue of tasks that have already been started and empty and cancel it when a new task comes in. The queue is emptied rather than removing the last element to handle multiple tasks being submitted at the same time and to handle if something cancels the task outside of this mechanism.