fix: properly handle stacker in wrapped tests
See, this is the part where for me it started to look like a problem. You know, I wanted to sacrifice the Precentral vein to order to get some exposure, but because of this guy's normal variation I got excited and all of a sudden I didn't know whether if I was looking at the Precentral vein or one of the internal cerebral veins or the vein of Galen or the Basilar vein of Rosenthal. So on my own, me, at this point I was ready to say, "That's it. Let's get out."