refactor(components,protocol-designer): Componentize StepContainer (#19241)
## Overview

There's a "Step Button" component in the Protocol Designer design
system, soon to be renamed "Step Container". We've been implementing it
with a "local" component in `protocol-designer/`, tied up with things
like Redux state, as opposed to a "real" component in `components/`.
This is a first step to making it a "real" component in `components/`
and unifying it with Figma.

Closes AUTH-1750.

## Changelog

This PR:

* Separates the presentational stuff from the more complex Redux state
stuff.
* Moves the presentational stuff into `components/` and makes a
Storybook story for it.

No functional change is intended.

Remaining known issues, which I'm not addressing in this PR:

* The overflow menu still appears when `size="iconOnly"`; the designs
imply it should go away. (But this would make some actions unreachable,
so I'm not sure it's intended in the designs. I've asked for
clarification in Figma.)
* We're missing the `focus-visible` styling as shown in the designs. I
view this as part of a more general thing where the component isn't
usable via keyboard overall. I think fixing this would involve a
reworking its DOM structure, like
https://stackoverflow.com/q/64522551/497934.
* The designs show the vertical size of the button shrinking when the
text is hidden, but it currently doesn't.
* It doesn't have a "subtext" line. (This is part of current work for
concurrent module actions, EXEC-1688).
* The designs show an error/warning icon separate from the step icon (so
2 icons total), but we currently have the error/warning icon *replace*
the step icon.
* It's still using styled-components instead of CSS Modules.