fix(app): Work around incomplete paints after a setup step has expanded (#19276)
## Overview

Closes RQA-4532. See that ticket for background.

## Changelog

* Whenever an expand/collapse animation completes, do some hacky shit to
trick the browser into repainting the screen. This seems to fix
RQA-4532. 🤷

* Change how the expand/collapse transition is computed.

Formerly, each section was a grid containing a single row, and we
transitioned that row's height between `0fr` and `1fr`. This is a known
trick for, effectively, transitioning between `height: 0` and `height:
auto`. But this did weird things at the intermediate stages of the
animation—there was empty space below the contents, like the contents
and container were growing/shrinking at different rates. The problem is
easier to see if you increase the transition time from 300 ms to like
3000 ms.

These days, we can just transition from `height: 0` to `height: auto`
directly, without grid tricks. So this PR does that to fix the empty
space problem.

* Refactor the way the `visibility` transition is specified. This is
just code cleanup, no user-visible effect.

`visibility` is all-or-nothing—it doesn't interpolate—so it was a little
misleading to use `ease` as the transition function. And I don't think
the duration of this transition needed to be different from the duration
of the height transition.