fix(api): Implement Flex Stacker deck conflict checking (#19260)
# Overview
Covers RQA-4539

The deck conflict checker is a circus and we play the role of it's
clowns.

We missed implementing the deck conflict checker for the Flex Stacker
module, which previously just returned None which proceeded to boot it
out of further deck conflict checking. This lead to issues like trash
bins being loadable into A3 when a stacker is in A4.

## Test Plan and Hands on Testing
- [x] Implement Unit Test to enforce stacker deck conflict check

- [x] The following protocol should fail analysis with this error:
"DeckConflictError [line 20]: flexStackerModuleV1 in slot B3 prevents
trash bin from using slot B3."
```
from opentrons.protocol_api import ProtocolContext

metadata = {
    "protocolName": "Flex Stacker Trashbin Error Protocol",
    "author": "Opentrons <protocols@opentrons.com>",
}
requirements = {
    "robotType": "Flex",
    "apiLevel": "2.25",
}
def run(protocol: ProtocolContext) -> None:
    """Protocol."""
    stacker_B = protocol.load_module( "flexStackerModuleV1", "B4")
    stacker_B.set_stored_labware(
        load_name="opentrons_flex_96_tiprack_200ul",
        count=1,
        lid="opentrons_flex_tiprack_lid",
    )

    trash = protocol.load_trash_bin("B3")

    tiprack = stacker_B.retrieve()
    protocol.move_lid(tiprack, trash, use_gripper=True)
```

## Changelog


## Review requests
So we wrap the Flex Stacker conflict errors in the "other module" class,
which is correct for now and kind of makes checking for it at all as
opposed to just returning under the "else" conditional redundant. Do we
want to leave it that way incase in the future we want a more fleshed
out stacker conflict error? Or should we just excise it entirely?

## Risk assessment
Low - fixes existing behavior

---------

Co-authored-by: Seth Foster <seth@opentrons.com>