fix(app): fix unhandled `go back` during compound pipette wizard flow (#19281)
Closes RQA-4534

The "Detach pipettes and attach 96 channel" flow is a special pipette wizard flow, because it is effectively a "compound" flow. We combine the "detach pipettes" flow with an "attach pipettes" flow.

When we "go back" on any step in any pipette flow, we return to the first step (unless we're on the last step). Going back to step 1 doesn't work in this particular compound flow if we are at an "attach" step, because we return to the initial "detach" step. Because those pipettes are no longer attached, internal state gets confused and we show nothing.

To fix, we should institute go back "checkpoints". As a user passes a checkpoint, pressing "go back" should return to the most recent checkpoint. So if a user presses "go back" during the attach portion of the flow, go back to the first attach step.