fix: remove unversioned liquid class changes (#19306)
These are good changes that are correct, but they are significant
behavior changes that must be separated by a new liquid class version.
They cannot retroactively change the V1 liquid class.

This was generated by running `git checkout
679ccd8ee5b409cce4c060641d46837de6aadf49 -- shared-data/liquid-class/`
which should give us the liquid classes as of
https://github.com/Opentrons/opentrons/commit/679ccd8ee5b409cce4c060641d46837de6aadf49
, the final commit before the problematic one.

The rest of https://github.com/Opentrons/opentrons/pull/19141 is
unchanged.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: sfoster1 <3091648+sfoster1@users.noreply.github.com>