refactor(api): use named constants for APIVersionError(until_version=...) (#19320)
# Overview

I noticed that we defined a constant for
`_LIQUID_CLASS_TRANSFER_TIP_RACKS_ARG_ADDED_IN = APIVersion(2, 25)`, but
then in the code, we write the version number as a literal:
```
raise APIVersionError(
    api_element="tip_racks",
    until_version="2.25",
)
```

In this PR, if there is a named constant for the `until_version`, I
changed the code to use that instead of a literal string.

We were also inconsistent in how we were stringifying the `APIVersion`
constant. Most places used `f"{VERSION_CONSTANT}"`, but some places used
`str(VERSION_CONSTANT)`. I changed them all to be consistent.

I also swapped the order of the `until_version` and `current_version`
args in a few places to be consistent.

## Test Plan and Hands on Testing

Run CI tests.

## Risk assessment

Low. Pure code cleanup with no behavioral change.