docs: temperature module manual migration (#19225)
# Overview

This migrates the temp mod manual to mkdocs.

RTC-782

## Test Plan and Hands on Testing

Review the docs for code and other errors.

## Changelog

Recent changes to the style:

- Removing logo from cover page. Title starts the doc.
- Moved WEEE info into compliance section, from warranty section.

## Review requests

Please review the code and writing for the usual.

## Risk assessment

Low, docs only.

<!--
- Indicate the level of attention this PR needs.
- Provide context to guide reviewers.
- Discuss trade-offs, coupling, and side effects.
- Look for the possibility, even if you think it's small, that your
change may affect some other part of the system.
- For instance, changing return tip behavior may also change the
behavior of labware calibration.
- How do your unit tests and on hands on testing mitigate this PR's
risks and the risk of future regressions?
- Especially in high risk PRs, explain how you know your testing is
enough.
-->