feat(api): Ensure Labware Schema 2 and 3 cross-compatibility (#19347)
Works towards EXEC-77

We want to support labware stackups that comprise of both labware schema 2 definitions and schema 3 definitions as best as possible, and this PR fills in the cross-compatibility gaps when support does not exist. Each parent-child labware stackup case is elaborated upon below.

Schema 3 Parent Definition, Schema 2 Child Definition

We are already fully covering this case. While the stackup integration snapshot testing covers testing for this currently, a650449 adds a unit test.

Schema 2 Parent Definition, Schema 3 Child Definition

Ok, this is the bigger issue, and the primary focus of this PR. When we place a schema 3 labware on top of a schema 2, we can't use locating features, so we have to fallback to the schema 2 way of doing things. That means we need to know the following:

* The cornerOffsetFromSlot in the child definition. This is easy and can be derived entirely from extents. Extents themselves are partially derived from cornerOffsetFromSlot as seen in the schema 2 to 3 migration script.
* The stackingOffsetWithLabware. Unfortunately, this is not easy...

After discussion with @SyntaxColoring, our plan is to provide affordances for specifying these offsets via a legacyStackingOffsetWithLabware field, but not include these values in any of our official definitions. The reason for this being that we'd have to maintain these values for every officially supported definition in perpetuity and constantly add to them as new labware gets released. In practice, it seems like there are only a few select group of users who could encounter this issue:

* Users that use a lower API Level than the Schema 3 release and then use Labware Creator to create a schema 3 labware. We can update Labware Creator with some warnings to mitigate this issue.
* Users that specify the version param when loading labware in protocols and mix schema 2 and 3 labware.
* Users that roll their own labware definitions (or with the help of Support).

So this seems like a case of choosing the better of two unideal scenarios.