refactor(api): calculate grip point from labware center position (#19354)
Closes EXEC-1751

When calculating the grip point, we currently use the center of the addressable area as our base position and adjust from there. This works currently, but it's more technically correct to grip from the expected labware bottom center position and apply offsets from that position. Doing so would in theory let us move labware that may not be in the center of an addressable area.