fix(protocol-designer): make volumeTooHigh() not crash when pipette is deleted (#19353)
# Overview

In PD, if you have a Mix step, then delete/replace the pipette, you'll
get a whitescreen crash with https://opentrons-76.sentry.io/issues/6805349834
RQA-4595.

The root cause is that when we try to regenerate the steps after the
pipette is deleted, the Mix step has a null pipette, and it crashes when
trying to call `volumeTooHigh()` to get the pipette capacity.

## Test Plan and Hands on Testing

Before this change, create a Mix step, then delete the pipette used for
the Mix. PD will crash.

Observe that PD no longer crashes after this change.

## Risk assessment

Low. Straightforward bug fix.