feat(api): Add ConcurrencyProvider to enable storage of state of concurrent commands (#19279)
# Overview

Add ConcurrencyProvider to enable storage of state of concurrent
commands.
Concurrency is needed for the robot to continue to new commands before
certain module commands are complete.

## Test Plan and Hands on Testing

There were two test files added.

`test_concurrency_provider`
This ensures that a concurrency provider makes a lock and that they
return the correct id. This is important because a lock keeps a task's
data from getting disrupted. Returning the correct id is needed for
referencing tasks later on in the protocol.

`test_task_handler`
This tests that tasks are run in the expected order and that the correct
tasks are cancelled. Tasks are cancelled by `group_id` which will be an
id that is associated to a unique module on the deck. When a task is
cancelled, it will only reference tasks in that group.

No hands on testing needed right now.

## Changelog
- Within the protocol engine command executor scaffolding was created to
enable module concurrent actions.
- `task_handler` holds rudimentary functions for task handling. 
- Currently, you can create a task, schedule running of more than one
task sequentially or concurrently, and cancel the latest task or the
previous task.
- `NOTE`: To cancel previous tasks, we keep a queue of tasks that have
already been started and empty and cancel it when a new task comes in.
The queue is emptied rather than removing the last element to handle
multiple tasks being submitted at the same time and to handle if
something cancels the task outside of this mechanism.
- In `./resources` a `concurrency_provider` was added to allow for keyed
lock storage for task concurrency. This is needed to ensure tasks do not
try to modify shared-data concurrently.

closes EXEC-1675