feat(api, shared-data): Support labware that does not follow parent rotation in labware schema 3 (#19314)
Works toward EXEC-77

In labware schema 2, we don't make affordances for labware that is asymmetric and do not follow the rotation of their parent deck item. In the current realm of Opentrons officially supported labware, there is only one labware that currently fits this criteria: the Opentrons Universal Flat Adapter. When placed on a heater shaker on the left side of the deck, the prongs face to the left. When placed on the right side of the deck, the prongs face to the right. So naturally, well A1 on the left side of the deck has a very different offset than well A1 on the right side of the deck.

To provide an adequate and accurate modeling of how this adapter mates with the heater shaker we have to account for the following physical realities:

* Some notion of deck "handedness". That is, we need to know which side of the deck the adapter resides on.
* The fact that labware residing on top of this adapter is always spring-pushed into the side of the adapter with the prongs (due to the Heater-Shaker springs).
* The fact that this adapter does not mate with the Heater-Shaker by simply being placed on top of the HS: it's held in place by a screw, and the adapter's pilot hole is not in the dead center of the labware.

Deck Handedness

We do this on the shared-data definition level.

* Waste chutes, fixed trashes, expansion slots, and base slots are not given an orientation. I'm using null to represent this in JSON, and this becomes the ModuleOrientation.NOT_APPLICABLE enum in Python.
* I've defined the magblock to always have a "center" orientation no matter the base slot. The alternative is we give this module the same treatment as base deck slots and give it a null orientation, but that seems less correct to me.
* Modules that can only go on one side of the deck are still given an orientation, since this is a physical truth about those modules. This is true of the Absorbance Reader and Flex Stacker.
* I didn't really know what orientation to give the TC, so I gave it a "left" orientation.
* In module defs, orientations are defined only for cutout ids that support that support the module.
* We can revisit these notions in the future, too. It's not a painful refactor to make if we deem it necessary.

Supporting labware residing on the Universal Flat Adapter

Relatively straightforward outside of the last commit. As outlined in the proposal, we want some sort of value to replace the stacking offsets, and when the contact surface of a labware is the external bottom of the wells instead of the bottom of the labware, we'd like to account for that. However, after syncing with Hardware, it turns out that suppliers don't always provide the number for the external bottom of wells to the bottom of labware (in this case, the opentrons_96_wellplate_200ul_pcr_full_skirt). The reason in part is due to factory differences on a batch basis and per-well basis. After some discussion, it sounds like the best thing to do for this specific labware is just assume there is no difference. Yes, it’s not ideal, but it sounds like we’re getting into the realm of minute differences that vary by labware batch, and given our plunger z-axis stepper can only handle a minimum value of 0.05mm, it seems we’re not dropping any egregious value.

But a few notes:

* We should still use this actual value when a supplier gives it to us.
* It’s probably worth exploring whether other manufacturing proxies for this value exist (although I’m skeptical).
* We should make sure we include this explanation in our final documentation when we roll out schema 3.

Screw Anchored Features

We do need to support this feature, because it reflects a reality of how labware can mate. This should hopefully be straightforward. There are affordances made for screwAnchoredAsChild despite the fact nothing in this PR makes use of it. It makes sense to get that in now though, because I propose that we do want to identify this as a real locating feature, and importantly, that the Universal Flat Adapter does not use this locating feature, because it can be placed only on top of the heater shaker. Other modules that contain screw anchoring support, such as the temp module, are not compatible with the Universal Flat Adapter.