refactor(protocol-designer): Use internal padding for the visual gap between step buttons (#19298)
## Overview

In the protocol timeline, there's a bit of space between adjacent step
buttons. In the code, we implement that with a flexbox gap. In the Figma
designs, it's implemented with an internal padding on each button.

Figma says we want to move towards the padding way:

> Design and dev decided to add 2px padding outside the colored
container area to eliminate the deck state flickering on hover when the
user’s mouse is in the 4px spacing between stepButtons.
> 
> This was defined at the start of the PD redesign, but was never
implemented into the component code-wise. We’d still like to pursue
adding this padding and thus will be leaving it into the [Figma]
component until the code catches up with design.

All the recent design work for concurrent module actions is based on the
padding way, so it seems like now is the time to change the code to work
like that too. Otherwise, we end up adding 2px adjustments everywhere
that don't appear in the designs, and things get pretty confusing.

Goes towards EXEC-1686 and EXEC-1688. (Not sure if there was ever a
ticket specifically for this?)

## Changelog

* Replace the flexbox gap with internal padding, as described above.
* Rename `StepContainer.tsx` to `ConnectedStepContainer.tsx`. I
accidentally left this out of #19241, sorry!