refactor(protocol-designer): make selectedLiquidId non-nullable in LiquidCardList/LiquidDetailCard (#19384)
# Overview

Another change for RQA-4612:

In `<LiquidCardList>`, we declared the `selectedLiquidId` prop as
optional, and in `<LiquidDetailCard>`, we declared the prop
`selectedValue` as optional.

However, these components are only ever rendered when there **IS** a
selected value:
https://github.com/Opentrons/opentrons/blob/37e527266d757b9d1111d907d1b60ed3ba049149/protocol-designer/src/components/organisms/SlotDetailModal/index.tsx#L171-L180

So we can get stronger type-checking in TypeScript by making
`selectedLiquidId` required instead of optional.

## Test Plan and Hands on Testing

Will run CI tests.

## Risk assessment

Low. Just a type declaration change, has no effect on functionality.