feat(api): add module asynch errors to hw callback
This lets the hardware controller dispatch notifications about
asynchronous errors to an engine, if one is running.

The engine, in turn, can decide whether or not to handle the error,
based on whether or not the module that encountered the error is
actually used in the protocol (this may be a decision we change our
minds about, or even make optional).

If the error is handled, the engine stops. This is similar in concept to
how we handle estop, but because the error will be coming from a module
and therefore won't interrupt engine execution, so we do have to make
the hardware stop going.