fix(api): don't touch-tip on source well after blowout to source if it's untouchable (#19425)
# Overview

This fixes AUTH-1871, where a customer was doing a liquid class transfer
with:
- A source that's a reservoir with `touchTipDisabled`.
- A destination that's a normal well plate.
- Blowout enabled after dispense with a location of `SOURCE`.
- Touch-tip enabled after dispense.

The customer was getting a `TouchTipDisabledError` when our
implementation tried to touch the reservoir.

In this fix, we disable the touch-tip after blowout when the blowout
location is `SOURCE` and the source has `touchTipDisabled`.

To keep this fix minimal, this PR does not change any behavior if the
blowout location is **not** `SOURCE`. For example, if a user is
transfering to a destination reservoir, and has touch-tip after dispense
enabled, we would continue to fail with `TouchTipDisabledError`. (In
that case, the user can disable touch-tip after dispense to get a
working transfer.)

## Test Plan and Hands on Testing

I analyzed the customer's protocol before and after this change, and
confirmed that we no longer fail with `TouchTipDisabledError` after this
change.

I also added a unit test where we try every combination of:
- transfer source is touchable or not
- transfer destination is a normal well or a trash bin
- the blowout location is `SOURCE`, `DESTINATION`, or `TRASH`

## Review requests

Sanniti's intent in the original implementation is a bit unclear.

In the original implementation of
`_do_touch_tip_and_air_gap_after_dispense()`, she had:
```
_do_touch_tip_and_air_gap_after_dispense():
    ...
    try:
        self._instrument.touch_tip(...)
    except TouchTipDisabledError:
        # TODO: log a warning
        pass
```
This implied that she intended to ignore any touch-tip failure after the
blowout. However, this code didn't work, because the exception that is
actually raised here is a `ProtocolCommandFailedError`.

There are other hints in the comments of the original implementation:

In `retract_after_dispensing()`, she wrote:
```
6. If blowout is “source” or “trash”
  - Move to location (top of Well)
  - Do blow-out (top of well)
  - Do touch-tip (?????) (only if it’s in a non-trash location)
```
so it looks like there was an open question of whether we wanted to do a
touch-tip after blowing out at all.

She also left the comment in `retract_after_dispensing()`
```
# TODO: no-op touch tip if touch tip is enabled and blowout is in trash/ reservoir/ any labware with touch-tip disabled
```
which implied that we shouldn't attempt to touch-tip at all if the
source is untouchable.


## Risk assessment

Low-ish.