feat(api): Consistently allow specifying namespaces and versions of adapters and lids (#19348)
## Overview

Closes the Python part of EXEC-1750.

The Python Protocol API has a number of "combination load methods" that
let you load multiple labware in a single call, like a main labware plus
an adapter, or a main labware plus a lid.

* `Labware`
  * `load_labware()`
* `ModuleContext`
  * `load_labware()`
  * `set_stored_labware()`
* `ProtocolContext`
  * `load_labware()`
  * `load_lid_stack()`

There are two things about these APIs today that are problematic:

1. There are no parameters to let you manually specify the namespace and
version of the adapter/lid, like you can for the main labware.

2. Without those parameters, the Python Protocol API tries to
automatically pick the namespace and version values for the adapter/lid.
The specific way that it does this is problematic. It will sometimes try
to match the the version of the adapter/lid to the version of the main
labware. But there's no reason to believe that that version of the
adapter/lid will even exist. For example, you could have a well plate
that goes up to v5 and a lid that only goes up to v2.

## Changelog

For all of the "combined load methods":

* Add new parameters:
  * `lid_namespace`
  * `lid_version`
  * `adapter_namespace`
  * `adapter_version`
* In the case where these parameters are not supplied, stop copying
their values from the namespace and version of the "main" labware.
Instead:`namespace` defaults to searching `opentrons` and custom
labware, and `version` defaults to something `apiLevel`-dependent. This
is consistent with how the basic case of
`ProtocolContext.load_labware(load_name, location)` works.
* Gate this behind ~~`"apiLevel": "2.25"`, which is the `apiLevel` being
released in robot software v8.6.0.~~, `"apiLevel": "2.26"`, which is the
next one in line to be released.
* Update docstrings enough to be correct, though perhaps not enough to
be good. I'll leave refinements to follow-up PRs.