feature(hardware-testing): add single tip pickup mode to 96 channel grav test (#19428)
# Overview
In order to support the way that hardware-testing handles LC testing, we
needed to add a single channel version of the grav script for the 96
channel.
This adds an option in the CSV to allow 96 channels to change from full
pickup to single pickup and use the vial as a test labware.