feat(api): leak check tool in analyze
Add a --leaks flag to the analysis CLI. This will cause the cli to print
an error if it detects that a protocol engine instance is still live
after the analysis has completed.

There's also a --leak-debug flag, which will drop you into a pdb
instance that lets you interactively poke at the engine object.