feat(api): add a leak check analysis cli (#19452)
We've realized that it's actually pretty easy for protocols with complex
code and functionality to accidentally leak a protocol engine. The long
term fix for this is to get protocols running in their own processes, or
at the very least their own interpreters; for now though, what we can do
is add a flag to the analysis cli that checks for the presence of such a
leak.

## Testing
- [x] leak an engine in a protocol and check that this prints it (I can
provide one that does this)
- [x] have a protocol that doesn't and see that it doens't print that

We should probably get this running automatically on our test harness.

---------

Co-authored-by: Josh McVey <josh.mcvey@opentrons.com>