refactor(api): connect module error callbacks
These can now source module errors and send them all the way up to the
engine and the robot server.

The module errors should only be triggered for significant hardware
errors that are detected by the module state pollers - condensation
fouling thermistors, yes; malformed gcode commands, no. They already
tend to force-disconnect the module.

The error handling behavior in general is not changed, other than that
the errors are now notified upwards as soon as we receive them.