feat(app, app-shell): support secondary windows in the desktop app (#19461)
Works toward EXEC-1760

With the camera enablement initiative, Design is implementing a new pattern utilizing "secondary windows." These are browser windows independent of the main browser window. We'll be using these for several items during camera enablement, such as the live stream and camera images, and because we will use these windows in the future, it makes sense to build a scalable architecture for adding secondary windows.

This commit adds the app-shell architecture for secondary windows. Secondary windows effectively have a "type", which is loosely associated to the kind of content the window renders. Each type of secondary window emits an action, and from that payload we build an id to determine the uniqueness of the content. If the content is unique, we open a new window, otherwise we refocus the window already displaying that content.

Because secondary windows do not require access to all action handlers, we only pass in a subset of them. Ex, we don't want update prompting to occur in a secondary window, so we don't need to supply the update handler to secondary windows.

The camera stream wire-up is added here for completeness and to provide a working example of the general interface.